CRUD Functionality for Admin Cats
Added edit, update, and destroy for making cats. Added necessary links, routes, and views. Added link to make a new cat in the admin dashboard and a link to the cat show page from the cat index. Had an issue when deleting a cat tied to an order, so added "dependent: :destroy" in the cat model to resolve issue.