Inactivate Cats
Admin can now change the status of a cat in the database to "inactive" (as opposed to deleting the cat outright).
Added custom route and action for changing status as well as a model method. Should not see "add to cart" on the cat index
if the cat's status is "inactive."